Simplify RoadGraph
Previously RoadGraph was modeling a graph where each tile either being a
node or being adjacent to a node was part of the graph. This was to
enable the path finder to find routes between two undevelopped zones
(which are not road tiles).

This moves the logic to connect an undevelopped zone to the road graph
in a separate class. When the connection is found, the path finding is
delegated to the A* path finder. This enables the RoadGraph to only
model roads (and drop connection to tiles adjacent to the road). This
also enables to cleanly separate the logic of 'how do I get from a
building to a road' in a separate class, from which we could decide to
alter it in the future (such as: every building < N tiles from the road
is connected).